# sdk/typescript/openapi.json
{"openapi":"3.0.1","info":{"title":"PALIMPSEST Engine API","description":"The ontology engine — sole write authority. Reads program against the generated TypeScript SDK. license_confirmed=false content is internal-use only.","license":{"name":"internal — see repository LICENSE decision"},"version":"v1"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/api/v1/import/batches":{"post":{"tags":["import-controller"],"operationId":"batch","parameters":[{"name":"kind","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeImportReportDto"}}}}}}},"/api/v1/admin/rebuild-projections":{"post":{"tags":["probe-controller"],"operationId":"rebuild","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/actions/undispute":{"post":{"tags":["action-controller"],"operationId":"undispute","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UndisputeReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/actions/supersede":{"post":{"tags":["action-controller"],"operationId":"supersede","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupersedeReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/actions/merge-entities":{"post":{"tags":["action-controller"],"operationId":"merge","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MergeReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/actions/dispute":{"post":{"tags":["action-controller"],"operationId":"dispute","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DisputeReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/actions/assert-claim":{"post":{"tags":["action-controller"],"operationId":"assertClaim","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JsonNode"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/actions/adjust-confidence":{"post":{"tags":["action-controller"],"operationId":"adjust","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdjustReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeMapStringObject"}}}}}}},"/api/v1/stats/summary":{"get":{"tags":["search-meta-controller"],"operationId":"stats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeStatsDto"}}}}}}},"/api/v1/search/entities":{"get":{"tags":["search-meta-controller"],"operationId":"search","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"cursor","in":"query","required":false,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListEntitySummaryDto"}}}}}}},"/api/v1/runs":{"get":{"tags":["search-meta-controller"],"operationId":"runs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListRunDto"}}}}}}},"/api/v1/runs/{id}":{"get":{"tags":["search-meta-controller"],"operationId":"run","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeRunDto"}}}}}}},"/api/v1/readyz":{"get":{"tags":["probe-controller"],"operationId":"readyz","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/v1/paths":{"get":{"tags":["export-events-path-controller"],"operationId":"paths","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"to","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"minConfidence","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"windowStart","in":"query","required":false,"schema":{"type":"string"}},{"name":"windowEnd","in":"query","required":false,"schema":{"type":"string"}},{"name":"temporalMode","in":"query","required":false,"schema":{"type":"string"}},{"name":"maxDepth","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":4}},{"name":"maxPaths","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":3}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopePathResult"}}}}}}},"/api/v1/meta/sources":{"get":{"tags":["search-meta-controller"],"operationId":"sources","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListSourceDto"}}}}}}},"/api/v1/meta/relation-types":{"get":{"tags":["search-meta-controller"],"operationId":"relationTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListRelationTypeDto"}}}}}}},"/api/v1/meta/agents":{"get":{"tags":["search-meta-controller"],"operationId":"agents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListAgentDto"}}}}}}},"/api/v1/healthz":{"get":{"tags":["probe-controller"],"operationId":"healthz","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/v1/export/claims":{"get":{"tags":["export-events-path-controller"],"operationId":"exportClaims","parameters":[{"name":"format","in":"query","required":false,"schema":{"type":"string","default":"ndjson"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/events":{"get":{"tags":["export-events-path-controller"],"operationId":"events","parameters":[{"name":"since","in":"query","required":false,"schema":{"type":"string"}},{"name":"cursor","in":"query","required":false,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListEventDto"}}}}}}},"/api/v1/entities/{id}":{"get":{"tags":["entity-controller"],"operationId":"view","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"resolution","in":"query","required":false,"schema":{"type":"string","default":"raw"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeEntityViewDto"}}}}}}},"/api/v1/entities/{id}/network":{"get":{"tags":["entity-controller"],"operationId":"network","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"minConfidence","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"windowStart","in":"query","required":false,"schema":{"type":"string"}},{"name":"windowEnd","in":"query","required":false,"schema":{"type":"string"}},{"name":"temporalMode","in":"query","required":false,"schema":{"type":"string"}},{"name":"includeUnscored","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeNetworkDto"}}}}}}},"/api/v1/entities/{id}/claims":{"get":{"tags":["entity-controller"],"operationId":"entityClaims","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"predicate","in":"query","required":false,"schema":{"type":"string"}},{"name":"role","in":"query","required":false,"schema":{"type":"string","default":"any"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"asserted"}},{"name":"minConfidence","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"includeUnscored","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"windowStart","in":"query","required":false,"schema":{"type":"string"}},{"name":"windowEnd","in":"query","required":false,"schema":{"type":"string"}},{"name":"cursor","in":"query","required":false,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeListClaimDetailDto"}}}}}}},"/api/v1/entities/{a}/relations/{b}":{"get":{"tags":["entity-controller"],"operationId":"pair","parameters":[{"name":"a","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"b","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"any"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopePairDossierDto"}}}}}}},"/api/v1/entities/random":{"get":{"tags":["entity-controller"],"operationId":"random","parameters":[{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"minScoredDegree","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeEntitySummaryDto"}}}}}}},"/api/v1/entities/lookup":{"get":{"tags":["entity-controller"],"operationId":"lookup","parameters":[{"name":"authority","in":"query","required":true,"schema":{"type":"string"}},{"name":"externalId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeEntitySummaryDto"}}}}}}},"/api/v1/claims/{id}":{"get":{"tags":["claim-controller"],"operationId":"claim","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeClaimDetailDto"}}}}}}},"/api/v1/claims/{id}/history":{"get":{"tags":["claim-controller"],"operationId":"history","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeHistoryDto"}}}}}}},"/api/v1/claims/{id}/evidence":{"get":{"tags":["claim-controller"],"operationId":"evidence","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvelopeEvidenceDto"}}}}}}}},"components":{"schemas":{"EnvelopeImportReportDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/ImportReportDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"ImportReportDto":{"type":"object","properties":{"run":{"type":"string"},"batch":{"type":"integer","format":"int32"},"received":{"type":"integer","format":"int32"},"inserted":{"type":"integer","format":"int32"},"duplicates":{"type":"integer","format":"int32"},"superseded":{"type":"integer","format":"int32"},"rejected":{"type":"array","items":{"$ref":"#/components/schemas/RejectDto"}}}},"Meta":{"type":"object","properties":{"requestId":{"type":"string"},"count":{"type":"integer","format":"int32"},"truncated":{"type":"boolean"},"counts":{"$ref":"#/components/schemas/NetworkCounts"},"page":{"$ref":"#/components/schemas/Page"},"asOfSystem":{"type":"string"},"statusFilter":{"type":"string"}}},"NetworkCounts":{"type":"object","properties":{"possibly":{"type":"integer","format":"int32"},"certainly":{"type":"integer","format":"int32"},"undated":{"type":"integer","format":"int32"}}},"Page":{"type":"object","properties":{"cursor":{"type":"string"},"limit":{"type":"integer","format":"int32"},"nextCursor":{"type":"string"}}},"RejectDto":{"type":"object","properties":{"line":{"type":"integer","format":"int32"},"reason":{"type":"string"}}},"EnvelopeMapStringObject":{"type":"object","properties":{"data":{"type":"object","additionalProperties":{"type":"object"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"UndisputeReq":{"type":"object","properties":{"claimId":{"type":"integer","format":"int64"}}},"JsonNode":{"type":"object"},"SupersedeReq":{"type":"object","properties":{"claimId":{"type":"integer","format":"int64"},"replacement":{"$ref":"#/components/schemas/JsonNode"},"reason":{"type":"string"}}},"MergeReq":{"type":"object","properties":{"memberEntityIds":{"type":"array","items":{"type":"integer","format":"int64"}},"rationale":{"type":"string"}}},"DisputeReq":{"type":"object","properties":{"claimId":{"type":"integer","format":"int64"},"ground":{"type":"string"},"reason":{"type":"string"}}},"AdjustReq":{"type":"object","properties":{"claimId":{"type":"integer","format":"int64"},"confidence":{"$ref":"#/components/schemas/JsonNode"},"reason":{"type":"string"}}},"EnvelopeStatsDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/StatsDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"HistogramBinDto":{"type":"object","properties":{"band":{"type":"string"},"count":{"type":"integer","format":"int32"}}},"SourceDto":{"type":"object","properties":{"slug":{"type":"string"},"title":{"type":"string"},"version":{"type":"string"},"retrievalUri":{"type":"string"},"license":{"type":"string"},"licenseConfirmed":{"type":"boolean"}}},"StatsDto":{"type":"object","properties":{"entities":{"type":"object","additionalProperties":{"type":"object"}},"claims":{"type":"object","additionalProperties":{"type":"object"}},"sourceRecords":{"type":"integer","format":"int32"},"sources":{"type":"array","items":{"$ref":"#/components/schemas/SourceDto"}},"confidenceHistogram":{"type":"array","items":{"$ref":"#/components/schemas/HistogramBinDto"}},"anomalyCounters":{"type":"object","additionalProperties":{"type":"object"}},"gender":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}},"temporalCodeShare":{"type":"object","additionalProperties":{"type":"object"}},"noRelationshipPct":{"type":"number","format":"double"}}},"EntitySummaryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"displayName":{"type":"string"},"entityType":{"type":"string"},"description":{"type":"string"},"gender":{"type":"string"},"lifeDates":{"$ref":"#/components/schemas/LifeDates"},"degreeScored":{"type":"integer","format":"int32"},"degreeUnscored":{"type":"integer","format":"int32"}}},"EnvelopeListEntitySummaryDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/EntitySummaryDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"LifeDates":{"type":"object","properties":{"bornEarliest":{"type":"string"},"bornLatest":{"type":"string"},"diedEarliest":{"type":"string"},"diedLatest":{"type":"string"}}},"EnvelopeListRunDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RunDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"RunDto":{"type":"object","properties":{"runId":{"type":"string"},"sourceSlug":{"type":"string"},"startedAt":{"type":"string"},"finishedAt":{"type":"string"},"batches":{"type":"integer","format":"int32"},"received":{"type":"integer","format":"int32"},"inserted":{"type":"integer","format":"int32"},"duplicates":{"type":"integer","format":"int32"},"superseded":{"type":"integer","format":"int32"},"rejected":{"type":"integer","format":"int32"},"manifest":{"type":"object"}}},"EnvelopeRunDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/RunDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"EnvelopePathResult":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/PathResult"},"meta":{"$ref":"#/components/schemas/Meta"}}},"PathResult":{"type":"object","properties":{"paths":{"type":"array","items":{"type":"array","items":{"$ref":"#/components/schemas/PathStep"}}},"budgetExceeded":{"type":"boolean"},"truncated":{"type":"boolean"}}},"PathStep":{"type":"object","properties":{"entityId":{"type":"integer","format":"int64"},"displayName":{"type":"string"},"claimId":{"type":"integer","format":"int64"},"confidence":{"type":"number","format":"double"}}},"EnvelopeListSourceDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/SourceDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"EnvelopeListRelationTypeDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RelationTypeDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"RelationTypeDto":{"type":"object","properties":{"slug":{"type":"string"},"label":{"type":"string"},"category":{"type":"string"},"rangeKind":{"type":"string"},"symmetric":{"type":"boolean"}}},"AgentDto":{"type":"object","properties":{"slug":{"type":"string"},"kind":{"type":"string"},"displayName":{"type":"string"}}},"EnvelopeListAgentDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AgentDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"EnvelopeListEventDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"EventDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"claimId":{"type":"integer","format":"int64"},"eventType":{"type":"string"},"actor":{"$ref":"#/components/schemas/AgentDto"},"occurredAt":{"type":"string"},"payload":{"type":"object"}}},"AttributeGroupDto":{"type":"object","properties":{"predicate":{"type":"string"},"claims":{"type":"array","items":{"$ref":"#/components/schemas/ClaimDetailDto"}}}},"ClaimDetailDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"subject":{"$ref":"#/components/schemas/EntitySummaryDto"},"predicate":{"type":"string"},"object":{"$ref":"#/components/schemas/ObjectDto"},"validTime":{"$ref":"#/components/schemas/FuzzyIntervalDto"},"confidence":{"$ref":"#/components/schemas/ConfidenceDto"},"method":{"type":"string"},"methodDetail":{"type":"object"},"status":{"type":"string"},"assertedBy":{"$ref":"#/components/schemas/AgentDto"},"recordedAt":{"type":"string"},"importRunId":{"type":"string"}}},"ConfidenceDto":{"type":"object","properties":{"effective":{"type":"number","format":"double"},"origin":{"type":"string"},"raw":{"type":"object"},"calibrated":{"type":"boolean"},"scored":{"type":"boolean"},"band":{"type":"string"}}},"CoverageDto":{"type":"object","properties":{"bySource":{"type":"array","items":{"$ref":"#/components/schemas/CoverageSourceDto"}},"scored":{"type":"integer","format":"int32"},"unscored":{"type":"integer","format":"int32"},"calibration":{"type":"string"}}},"CoverageSourceDto":{"type":"object","properties":{"slug":{"type":"string"},"relationshipClaims":{"type":"integer","format":"int32"},"attributeClaims":{"type":"integer","format":"int32"}}},"EntityViewDto":{"type":"object","properties":{"summary":{"$ref":"#/components/schemas/EntitySummaryDto"},"externalIds":{"type":"array","items":{"$ref":"#/components/schemas/ExternalIdDto"}},"attributeClaims":{"type":"array","items":{"$ref":"#/components/schemas/AttributeGroupDto"}},"coverage":{"$ref":"#/components/schemas/CoverageDto"}}},"EnvelopeEntityViewDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/EntityViewDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"ExternalIdDto":{"type":"object","properties":{"authority":{"type":"string"},"externalId":{"type":"string"}}},"FuzzyIntervalDto":{"type":"object","properties":{"startEarliest":{"type":"string"},"startLatest":{"type":"string"},"endEarliest":{"type":"string"},"endLatest":{"type":"string"},"approximate":{"type":"boolean"},"original":{"type":"object"}}},"LiteralDto":{"type":"object","properties":{"kind":{"type":"string"},"value":{"type":"object"},"authority":{"type":"string"}}},"ObjectDto":{"type":"object","properties":{"entity":{"$ref":"#/components/schemas/EntitySummaryDto"},"literal":{"$ref":"#/components/schemas/LiteralDto"}}},"EdgeDto":{"type":"object","properties":{"claimId":{"type":"integer","format":"int64"},"counterpart":{"$ref":"#/components/schemas/EntitySummaryDto"},"predicate":{"type":"string"},"confidence":{"$ref":"#/components/schemas/ConfidenceDto"},"validTime":{"$ref":"#/components/schemas/FuzzyIntervalDto"},"scored":{"type":"boolean"},"certainlyActive":{"type":"boolean"},"undated":{"type":"boolean"},"status":{"type":"string"}}},"EnvelopeNetworkDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/NetworkDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"NetworkDto":{"type":"object","properties":{"focus":{"$ref":"#/components/schemas/EntitySummaryDto"},"edges":{"type":"array","items":{"$ref":"#/components/schemas/EdgeDto"}}}},"EnvelopeListClaimDetailDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ClaimDetailDto"}},"meta":{"$ref":"#/components/schemas/Meta"}}},"EnvelopePairDossierDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/PairDossierDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"PairDossierDto":{"type":"object","properties":{"a":{"$ref":"#/components/schemas/EntitySummaryDto"},"b":{"$ref":"#/components/schemas/EntitySummaryDto"},"claims":{"type":"array","items":{"$ref":"#/components/schemas/ClaimDetailDto"}}}},"EnvelopeEntitySummaryDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/EntitySummaryDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"EnvelopeClaimDetailDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/ClaimDetailDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"EnvelopeHistoryDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/HistoryDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"HistoryDto":{"type":"object","properties":{"events":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}}},"EnvelopeEvidenceDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/EvidenceDto"},"meta":{"$ref":"#/components/schemas/Meta"}}},"EvidenceDto":{"type":"object","properties":{"claim":{"$ref":"#/components/schemas/ClaimDetailDto"},"support":{"type":"array","items":{"$ref":"#/components/schemas/SupportDto"}}}},"SourceRecordDto":{"type":"object","properties":{"recordKind":{"type":"string"},"externalId":{"type":"string"},"contentHash":{"type":"string"},"raw":{"type":"object"}}},"SupportDto":{"type":"object","properties":{"source":{"$ref":"#/components/schemas/SourceDto"},"record":{"$ref":"#/components/schemas/SourceRecordDto"}}}}}}